New Backbone: YoloWorld!

+++ sgg_benchmark/modeling/roi_heads/relation_head/loss/loss.py
@@ -85,7 +85,7 @@ class RelationLossComputation(object):
         fg_labels = cat([proposal.get_field("labels") for proposal in proposals], dim=0)
         rel_labels = cat(rel_labels, dim=0)
 
-        loss_relation = self.criterion_loss(relation_logits, rel_labels.long())
+        loss_relation = self.criterion_loss(relation_logits, rel_labels)
 
         if refine_obj_logits is not None:
             refine_obj_logits = cat(refine_obj_logits, dim=0)
